// pslpwcl/src/app/modules/pslpapi/model/collegamentoAnnuncioCandidaturaRequest.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CallInfo } from './callInfo';
import { VacancyCandid } from './vacancyCandid';

export interface CollegamentoAnnuncioCandidaturaRequest { 
    callInfo: CallInfo;
    dataVisibilitaVacancyCandid?: Date;
    flgInvioMassivoMailAzienda?: boolean;
    flgInvioMassivoMailCittadino?: boolean;
    idMotivoVacancyCandid?: number;
    idStatoVacancyCandidDaImpostare?: number;
    idVacancyCandidList?: Array<number>;
    note?: string;
    vacancyCandid: VacancyCandid;
}