// pslpwcl/src/app/modules/pslpapi/model/vacancyCandid.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CpiBlp } from './cpiBlp';
import { DatiAnnuncio } from './datiAnnuncio';
import { DatiCandidato } from './datiCandidato';
import { VacancyCandidStato } from './vacancyCandidStato';

export interface VacancyCandid { 
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dInserim: Date;
    d_vacancy_candidatura: Date;
    d_visibilita_vacancy_candid?: Date;
    data_ora_colloquio?: Date;
    datiAnnuncio?: DatiAnnuncio;
    datiCandidato?: DatiCandidato;
    flg_auto_candidatura?: string;
    flg_invitato?: string;
    flg_suggerito_operatore?: string;
    idVacancyCandid: number;
    id_candidatura?: number;
    id_cpi_operatore?: CpiBlp;
    id_operatore?: string;
    id_sil_lav_anagrafica?: number;
    id_vacancy: number;
    vacancyCandidStatoList?: Array<VacancyCandidStato>;
}